Fixed issues with game state serialization.
