+++ game.hpp
@@ -163,6 +163,8 @@ public:
     uint32_t length;
     char direction;
 
+    uint32_t id;
+
     Snake(float x, float y) {
         squares = std::vector<Square*>();
         for (int i = 0; i < init_snake_len; i++) {
@@ -214,7 +216,6 @@ private:
 
     Snake* snake;
     Apple* apple;
-    struct game_state state;
 
     pthread_t listener_thrd, sender_thrd;
 
@@ -224,79 +225,12 @@ private:
 
     void process_input(GLFWwindow* window);
 
+    void update_game_state(Snake* snake, Apple* apple);
+
 public:
-    Game() {
-        glfwInit();
-        glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
-        glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 3);
-        glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
+    Game();
 
-        window = glfwCreateWindow(800, 800, "Snake", NULL, NULL);
-        if (window == NULL) {
-            std::cout << "ERROR: Failed to create a window\n";
-            glfwTerminate();
-            exit(1);
-        }
-
-        glfwMakeContextCurrent(window);
-        if (!gladLoadGLLoader((GLADloadproc)glfwGetProcAddress)) {
-            std::cout << "Failed to initialize GLAD\n";
-            glfwTerminate();
-            exit(1);
-        }
-
-
-        glViewport(0, 0, 800, 800);
-        glfwSetFramebufferSizeCallback(window, framebuffer_size_callback);
-
-
-        std::string vert_shader_src = load_shader_src(VERT_SHADER_PATH);
-        std::string frag_shader_src = load_shader_src(FRAG_SHADER_PATH);
-
-        const char* vert_shader_src_cstr = vert_shader_src.c_str();
-        const char* frag_shader_src_cstr = frag_shader_src.c_str();
-
-        GLuint vert_shader = glCreateShader(GL_VERTEX_SHADER);
-        GLuint frag_shader = glCreateShader(GL_FRAGMENT_SHADER);
-
-        glShaderSource(vert_shader, 1, &vert_shader_src_cstr, NULL);
-        glShaderSource(frag_shader, 1, &frag_shader_src_cstr, NULL);
-
-        glCompileShader(vert_shader);
-        glCompileShader(frag_shader);
-
-        if (check_shader_error(vert_shader)) {
-            glfwTerminate();
-            exit(1);
-        }
-        if (check_shader_error(frag_shader)) {
-            glfwTerminate();
-            exit(1);
-        }
-
-        shader_program = glCreateProgram();
-        glAttachShader(shader_program, vert_shader);
-        glAttachShader(shader_program, frag_shader);
-        glLinkProgram(shader_program);
-        glDeleteShader(vert_shader);
-        glDeleteShader(frag_shader);
-
-        snake = new Snake(0.0f, 0.0f);
-
-        snake->num_buffers = snake->length;
-        snake->VAOs = std::vector<GLuint>(snake->num_buffers);
-        snake->VBOs = std::vector<GLuint>(snake->num_buffers);
-        snake->EBOs = std::vector<GLuint>(snake->num_buffers);
-
-        snake->gen_vertex_objs(snake->VAOs, snake->VBOs, snake->EBOs);
-        apple = new Apple(Apple::rand_float(), Apple::rand_float());
-
-        connect_to_server();
-    }
-
-    ~Game() {
-        glfwTerminate();
-    }
+    ~Game() { glfwTerminate(); }
 
     std::string load_shader_src(std::string path);
 
@@ -314,9 +248,9 @@ public:
 
     static void* sender(void* args);
 
-    std::string serialize_state(struct game_state state);
+    static std::string serialize_state(struct game_state state);
 
-    struct game_state deserialize_state(std::string json_data);
+    static struct game_state deserialize_state(std::string json_data);
 
     // Writes the current game state to the server
     int write_state();
